rename python support lib
