+++ src/danfo/index.ts
@@ -5,7 +5,7 @@ import {
   JupyterFrontEnd,
   JupyterFrontEndPlugin
 } from '@jupyterlab/application';
-import { ITLabPyDSManager } from '../store-python/datasource';
+import { ITLabPyDSManager } from '../python/datasource';
 import { ITLabStoreManager } from '../store/manager';
 import { danfoDFModel } from './model';
 import { danfoPythonDS } from './python';
