+++ src/index.ts
@@ -4,8 +4,8 @@
 import { labDanfoPlugin } from './danfo';
 import { labFrontManagerPlugin } from './front';
 import { labPrimitivesPlugin } from './primitives';
+import { labStorePythonPlugin } from './python';
 import { labStoreManagerPlugin } from './store';
-import { labStorePythonPlugin } from './store-python';
 import { labWidgetPlotlyPlugin } from './widget-plotly';
 import { labWidgetStorePlugin } from './widget-store';
 
@@ -16,8 +16,8 @@ const plugins = [
   labDanfoPlugin,
   labFrontManagerPlugin,
   labPrimitivesPlugin,
-  labStoreManagerPlugin,
   labStorePythonPlugin,
+  labStoreManagerPlugin,
   labWidgetPlotlyPlugin,
   labWidgetStorePlugin
 ];
