+++ src/primitives/python.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2022, twiinIT
 // SPDX-License-Identifier: BSD-3-Clause
 
-import { IPyDataSource } from '../store-python/datasource';
+import { IPyDataSource } from '../python/datasource';
 
 const nullPythonDS: IPyDataSource = {
   id: 'primitives_null',
